+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - The barebone ALE function `.ale()` has become faster thanks to [issue #11](https://github.com/mayer79/effectplots/issues/11) by [@SebKrantz](https://github.com/SebKrantz).
 - Subsampling indices for outlier capping is now done only once, instead of once per feature.
+- Speed-up for discrete numeric features.
 
 ## Bug fixes